Move networking to a separate thread
This rework splits the networking system to a separate crate and moves the processing to a networking thread. The main thread will no longer be blocked by networking operations, which was the main reason for the client freezing.

I took this opportunity to also refactor the way that networking routines are handled. Instead of sending a packet and waiting for a response, the request and response are completely decoupled, offering more resilience to packets arriving out of order.