Rework FromBytes, ToBytes, ByteConvertable, and Packet derive macros
Previously, a lot of types could only be deserialized but not
serialized. The goal is to make it easier to convert both ways, and to
make instantiation of `ToBytes` types easier by deriving an associated `new` function that initializes fields that have a fixed value or a value that can be derived.