Add point lights with drop shadows
This commit adds the MVP of drop shadows for point lights. This includes
a fixed amount of point lights that can be rendered with shadows and a
point light manager that takes care of deciding which point light
shadows are most important for the scene.

Additionally it improves the way frustum culling works to play better
with point lights.

Also includes small unrelated cleanups.