Rework main application loop.
We used the old pattern of handling the windows and surface lifetimes, which is not the most portable way do it. winit recommends this pattern, so that we can properly support for example Android, iOS and the web.

This change also has the added benefit, that we start to separate the gigantic main function into smaller, easier to understand parts.

I had to change some state tracking, so that it could properly be defined as a struct (anonymous closure don't really have a valid function signature you could write down, so I had to create a reference counted trait object).