Picker values can fully store 32-bit values.
This removes the implicit reliance of an entity ID to not use the full 32-bit value range. Since we can't assume the actual range that is used inside the 32 bit of an EntityId, we need to properly encode the full 32 bit inside the picker buffer.

For this we use a Rg32Uint texture, to store the picker values.
This doubles the picker texture size (4K texture is around 65 MiB).
The runtime of the shader is still very low and insignificant compared to other render passes.

This also allows us to encode much more data into the picker value later, if we need to do so.