Make graphics settings configurable
I included a hard copy of a frame pacer that I used in the past (it's not published). It is licensed under zlib license. The inclusion of the header is enough to fulfill the license requirements. No need to include it later inside the binary release.

For people that want the best possible latency without running their CPU hot, I included said frame pacer to limit the user selectable framerate.