Even faster way to do cone/sphere intersections
This is now optimized to the extreme. We still have
one 3x3 matrix operation in the hot loop, but modern
GPU have fast paths for those. The intersection test
itself is now very, very efficient. I doubt there is
something more efficient out there. We take full
advantage of the fact, that our cone is axis aligned
and also the origin of the space.