Use signed distance to render smooth edges
First of all, I removed MSAA on the interface buffer. MSAA doesn't really help with the interface, because we essentially only render rectangles with no real edges.

We instead will now always try to use analytical antialiasing where ever possible.

Currently, we need smoothing for the edges of rectangles. This is now done by using a signed distance based approach, which mathematically defines the properties of our rectangle with rounded edges. We then use the signed distance to properly antialias the edges, using the screen space derivative and the smoothstep function.

I also tried to fix visual defects that became visible because of the lack of MSAA (profiler).