Add multi-sampling of the edges and HQ UI mode
Even though the edges should be pretty much perfect mathematically speaking, in reality there are small imperfections, when doing a simple smoothstep approach. This is mainly because of the way GPU do the rasterization.

We now brute-force the antialiasing by applying x16 MSAA inside the shader, only at the borders of a rectangle. The performance impact is very small, since no texture lookups are done and the "sampling" is done purely mathematically.

This PR also adds a HQ mode for the UI, that allows the user to run the UI in double the screen resolution.