Implement and use premultiplied alpha
Since this fixes one of the issues with the white border around the entities, we should switch to pre-multiplied alpha throughout the render pipeline.

I did my best to catch all places where we had to change the logic when handling alpha values.

I moved the premultiply_alpha() function into the util crate, so that it gets properly optimized when compiling debug builds.