Implement proper text shaping
This replaces rusttypes with cosmic_text, which
implements a layout engine and proper text shaping.

We don't use it to the fullest yet, but that would require a redesign of how we handle text in our interface.

I changed the sampler for text to nearest, which results in much shaper text.