Only create sub-meshes of transparent vertices
Some nodes have a mix of opaque and transparent vertices. That means we have to extract the transparent vertices from them and then only create sub meshed from those transparent vertices.

I also increased the EPSILON value, so that we only create sub-meshes if the vertices of them are clearly seperated in model space.