Remove CMAA2 to properly render effects again
We switched to a float based render target for CMAA2, but effects need a render target that is clamped to u8, so we have to remove CMAA2 to get proper effect rendering again.

This also reverts back the render target to use Rgba8UnormSrgb, and effects to use straight alpha, so that all effect are rendered properly again.