Fix bug in frustum creation code
The near and far planes where "switched" when we used reverse Z projection. It didn't matter for actual frustum culling (since the order is not important), but we should have always the correct order anyhow. We also detect now, when we use an infinite projection and properly set the normal vector (it would be NaN otherwise).