Render in-game SDF sprites with a black outline
SDF sprites are created with a pxrange of 8 (4 inside, 4 outside pixel range) and a size of 64x64. SDF are also now always grayscale and are mainly colored using the instruction color. For this we extended the texture loader to also handle grayscale images properly. Only PNG and TGA provide proper grayscale images.