Implement MSDF font rendering and change default font
We now use Noto Sans, since it has the broadest Unicode support of the open source fonts. I added a README.md in the font folder, so that people can change the font if the need should arise to do so.

With this change, we unified our text rendering to use MSDF rendering. For this we have to generate the whole font map offline in advance.