Implement custom path finding
I found the path finding code highly complicated and also inefficient, since we allocated on the heap multiple times, without amortising the allocations.

This implementation provides a struct that implements the A* and also line of sight path finding in a way, that is compatible with the server implementations. We also re-use the heap allocations.

The A* algorithm is a basic implementation that uses a binary heap. Moved the code in korangar_utils, so that it can be properly optimized and also tested.