Separate sprite and entity animations
This PR separates the action types and the way we render the animations of entities and "actor" sprites like the mouse cursor and the skill icons completely. We also now properly calculate the base offset of entity actions based on the entity type. The workaround for the mouse cursor has also been removed by fixing the underlying problem.