Improve first map server connect
On first login the map server will send a command to change the map regardless, so we don't need to handle the map loading in the network event when we disconnect from the character server. We only initialize the player entity and interface in this case. We also don't use the position that the map server sends us in the MapServerLoginSuccessPacket. The ChangeMapPacket is our only signal that we use to first load the new map and then set the correct position. The three positions that we used before this PR lead to some desync in the first few seconds, since it seems, only the final position of the map change network event is the correct one.

We also now properly turn the screen black until we fully finished loading the map.

Later we should also do the loading asynchronously on map change and turn the screen black while loading.