Use zlib-rs as primary deflate backend
This fixes the issue we had with the "data\wav\worm_tail_damage.wav" sound file that can't be read with miniz_oxide of libflate.

Interestingly, the trait BufRead is implemented for `&[u8]` and the internals of `read_to_end()` essentially handles the writing into uninitialized memory of the Vector extremely efficiently then.