Support texture compression
This uses the block_compression crate that I wrote.

It is using WGPU compute shader to compress the textures using BC7 on the fly when loading the map.

Textures using BC7 are 1/4 of the uncompressed texture size. Quality is defined by "speed" like video encoders do: the slower, the better the quality.