Remove glidesort
Glidesort is actually unmaintained, since it's successor, driftsort, is now the current implementation in std. Since we don't want to allocate each sort/frame, we use unstable sort again. For models, we now remove the z-fighting by using a very small offset for each node in a model.

Entities might fight, but this is very rare and should not be discernible, by the normal "z-fighting" that occurs by the normal entity movement.