Implement WBOIT
We now support order independent transparency based on "Weighted Blended Order-Independent Transparency" as designed by Morgan McGuire and Louis Bavoil.

The weight function used is designed for depth values between 0 and 500. This means, that transparent objects are rendered correctly, if they are between 0 and 500 world space units in from of the camera.