Implement texture atlas cache
The cache can be generated by calling korangar with "sync-cache". It takes around 30 to 60 minutes for the first run. If the game files are updated, the cache can be re-synched by calling korangar with "sync-cache" again, which will then only takes around 3 minutes to check all texture atlases. We use blake3 for our checksums, since it's extremely fast.