Improve performance and system responsiveness of sync-cache
We do the checking in parallel with rayon and submit the compression tasks in smaller chunks.

I removed the debug code and instead used printf directly, since it's not advices to use the debug feature while running this command (since we produce a log of debug output at the same time).