Rework model rendering
The game file models are created in a way, that they need proper wrapping texture sampling. This is not 100% correctly possible with our texture atlas. This is especially the case for the RSM 2 models, which have animated textures, that rely on texture wrapping. We instead now use simpler texture binding arrays instead.

This is well support supported on desktop and only Metal currently uses the non-bindless code path, but this will be fixed with the next WGPU release.

For the non-bindless path we split the meshes into sub-meshes that use the same texture of a model.

Cache should be re-synched, since we now simply create compressed version of all textures. Takes longer then previously, but load time are good now and we can re-use textures between different map loads.