Fix pop sound on linux on startup
The maintainer of Kira found the root cause of the pop sound under linux on first start:

The initialized the ambient listener's rotation with a "zero" quaternion, which led to NAN values. We instead should initialize it to any other value, so that no NAN values are generated using it.

The previous workaround is not needed anymore.