Change video codec to AV1
h264 still has open patents, so it's safer to use AV1.

This would only be an issue if we would distribute the client as a binary in the future. But using AV1 is the safest option.

Another benefit: rav1d is a native Rust implementation.

Only downside: `nasm` needs to be installed when building.