Rework cache configurations
I currently think that actions, animations and effects are better for us to cache by count and not take their size into consideration. Textures, sprites and sounds are still limited by both. We should add a debug windows to show the current cache sizes and also make the amount to cache configurable.