Fixe the S3-FIFO bug
The S3-FIFO algorithm is wrongly documented in the official and current version of the paper. There were no errata, but on a blog post there was a sidenote about this issue. Very frustrating.

The old code did essentially promote the whole small FIFO once the small FIFO got full. This change makes sure that there is only one promotion or eviction in the small FIFO.