Reduce shadow memory usage and better clamp planes
Using Depth16Unorm uses only half of the memory for the directional shadow map. This improves performance especially with soft shadows. I clamped the shadows a bit more, so that we get some more precision. But the current precision is more than fine for the artwork and viewpoint of RO.